organized
yay i organized my code
